requirements.txt
openai
python-dotenv
pdfplumber
fastapi "uvicorn[standard]" 
python-multipart